+++ docusaurus.config.js
@@ -88,6 +88,7 @@ module.exports = {
             darkTheme: require("prism-react-renderer/themes/vsDark"),
         },
         algolia: {
+            appId: 'BH4D9OD16A',
             apiKey: "c0cf59beed38709e9ed6b0ac80b24ee5",
             indexName: "typescripttolua",
         },
@@ -109,5 +110,8 @@ module.exports = {
             },
         ],
     ],
-    plugins: [require.resolve("./docusaurus-plugin")],
+    plugins: [
+        "docusaurus-plugin-sass",
+        require.resolve("./docusaurus-plugin")
+    ],
 };
