+++ package.json
@@ -27,28 +27,29 @@
         "fengari-web": "^0.1.4",
         "lua-types": "^2.11.0",
         "lz-string": "^1.4.4",
-        "monaco-editor": "^0.25.2",
+        "monaco-editor": "^0.34.0",
         "path-browserify": "^1.0.1",
         "process": "^0.11.10",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
-        "react-json-tree": "^0.15.0",
-        "react-monaco-editor": "^0.43.0",
+        "react-json-tree": "^0.17.0",
+        "react-monaco-editor": "^0.50.1",
         "stream-browserify": "^3.0.0",
         "typescript-to-lua": "^1.9.0"
     },
     "devDependencies": {
         "@ark120202/typescript-config": "^2.2.0",
-        "@docusaurus/core": "^2.0.0-beta.0",
-        "@docusaurus/preset-classic": "^2.0.0-beta.0",
+        "@docusaurus/core": "^2.1.0",
+        "@docusaurus/preset-classic": "^2.1.0",
         "@types/node": "^14.14.2",
+        "docusaurus-plugin-sass": "^0.2.2",
         "file-loader": "^6.2.0",
         "fork-ts-checker-webpack-plugin": "^6.2.10",
         "null-loader": "^4.0.1",
-        "prettier": "2.3.1",
+        "prettier": "^2.7.1",
         "raw-loader": "^4.0.2",
-        "sass": "^1.27.0",
-        "sass-loader": "^12.1.0",
+        "sass": "^1.54.8",
+        "sass-loader": "^13.0.2",
         "typescript": "^4.7.4",
         "worker-loader": "^3.0.8"
     },
